requirements.txt
RPi.GPIO==0.6.3
satori-rtm-sdk==1.3.0
git+https://github.com/adafruit/Adafruit_Python_DHT
git+https://github.com/sunfounder/Adafruit_Python_BMP
smbus2==0.2.0